Rename root package to AD
