+++ AD/MVFA/CLIP/__init__.py
@@ -1,0 +1,2 @@
+# CLIP submodule package
+
