+++ AD/MVFA/CLIP/adapter.py
