+++ AD/MVFA/CLIP/clip.py
