+++ AD/MVFA/CLIP/model.py
