+++ AD/MVFA/CLIP/model_configs/ViT-L-14-336.json
