+++ AD/MVFA/CLIP/modified_resnet.py
