+++ AD/MVFA/CLIP/openai.py
