+++ AD/MVFA/LICENSE
