+++ AD/MVFA/dataset/fewshot_seed/Retina_RESC/8-shot.txt
