+++ AD/MVFA/test_zero.py
@@ -9,15 +9,15 @@ from torch.nn import functional as F
 from tqdm import tqdm
 from sklearn.metrics import roc_auc_score
 from scipy.ndimage import gaussian_filter
-from dataset.medical_zero import MedTestDataset, MedTrainDataset
-from CLIP.clip import create_model
-from CLIP.tokenizer import tokenize
-from CLIP.adapter import CLIP_Inplanted
+from AD.MVFA.dataset.medical_zero import MedTestDataset, MedTrainDataset
+from AD.MVFA.CLIP.clip import create_model
+from AD.MVFA.CLIP.tokenizer import tokenize
+from AD.MVFA.CLIP.adapter import CLIP_Inplanted
 from PIL import Image
 from sklearn.metrics import precision_recall_curve
-from loss import FocalLoss, BinaryDiceLoss
-from utils import augment, encode_text_with_prompt_ensemble
-from prompt import REAL_NAME
+from AD.MVFA.loss import FocalLoss, BinaryDiceLoss
+from AD.MVFA.utils import augment, encode_text_with_prompt_ensemble
+from AD.MVFA.prompt import REAL_NAME
 
 import warnings
 warnings.filterwarnings("ignore")
@@ -46,6 +46,7 @@ def main():
     parser.add_argument('--data_path', type=str, default='./data/')
     parser.add_argument('--batch_size', type=int, default=1)
     parser.add_argument('--img_size', type=int, default=240)
+    parser.add_argument('--save_path', type=str, default='./ckpt/zero-shot/')
     parser.add_argument("--epoch", type=int, default=50, help="epochs")
     parser.add_argument("--learning_rate", type=float, default=0.0001, help="learning rate")
     parser.add_argument("--features_list", type=int, nargs="+", default=[6, 12, 18, 24], help="features used")
@@ -61,8 +62,9 @@ def main():
     model = CLIP_Inplanted(clip_model=clip_model, features=args.features_list).to(device)
     model.eval()
 
-    for name, param in model.named_parameters():
-        param.requires_grad = True
+    checkpoint = torch.load(os.path.join(f'{args.save_path}', f'{args.obj}.pth'))
+    model.seg_adapters.load_state_dict(checkpoint["seg_adapters"])
+    model.det_adapters.load_state_dict(checkpoint["det_adapters"])
 
 
     # optimizer for only adapters
@@ -92,86 +94,8 @@ def main():
             text_feature = encode_text_with_prompt_ensemble(clip_model, REAL_NAME[CLASS_INDEX_INV[i]], device)
             text_feature_list.append(text_feature)
 
-    save_score = 0.0
-
-    for epoch in range(args.epoch):
-        print('epoch', epoch, ':')
-
-        loss_list = []
-        idx = 0
-        for (image, image_label, mask, seg_idx) in tqdm(train_loader):
-            if idx % (len(train_loader) // 5) == 0:
-                score = test(args, model, test_loader, text_feature_list[CLASS_INDEX[args.obj]])
-                if score >= save_score:
-                    save_score = score
-                    ckp_path = f'./ckpt/zero-shot/{args.obj}.pth'
-                    torch.save({'seg_adapters': model.seg_adapters.state_dict(),
-                                'det_adapters': model.det_adapters.state_dict()}, 
-                                ckp_path)
-                    print(f'best epoch found: epoch {epoch} batch {idx}')
-                print('\n')
-            idx += 1
-
-            image = image.squeeze(0).to(device)
-            seg_idx = seg_idx.item()
-
-            with torch.cuda.amp.autocast():
-                _, seg_patch_tokens, det_patch_tokens = model(image)
-                seg_patch_tokens = [p[0, 1:, :] for p in seg_patch_tokens]
-                det_patch_tokens = [p[0, 1:, :] for p in det_patch_tokens]
-
-                # image level
-                det_loss = 0
-                image_label = image_label.squeeze(0).to(device)
-                for layer in range(len(det_patch_tokens)):
-                    det_patch_tokens[layer] = det_patch_tokens[layer] / det_patch_tokens[layer].norm(dim=-1, keepdim=True)
-                    anomaly_map = (100.0 * det_patch_tokens[layer] @ text_feature_list[seg_idx]).unsqueeze(0)    
-                    anomaly_map = torch.softmax(anomaly_map, dim=-1)[:, :, 1]
-                    anomaly_score = torch.mean(anomaly_map, dim=-1)
-                    det_loss += loss_bce(anomaly_score, image_label)
-
-
-                if seg_idx > 0:
-                    # pixel level
-                    seg_loss = 0
-                    mask = mask.squeeze(0).to(device)
-                    mask[mask > 0.5], mask[mask <= 0.5] = 1, 0
-                    for layer in range(len(seg_patch_tokens)):
-                        seg_patch_tokens[layer] = seg_patch_tokens[layer] / seg_patch_tokens[layer].norm(dim=-1, keepdim=True)
-                        # print(seg_patch_tokens[layer].shape, text_feature_list[seg_idx].shape) # torch.Size([289, 768]) torch.Size([768, 2])
-                        anomaly_map = (100.0 * seg_patch_tokens[layer] @ text_feature_list[seg_idx]).unsqueeze(0)
-                        B, L, C = anomaly_map.shape
-                        H = int(np.sqrt(L))
-                        anomaly_map = F.interpolate(anomaly_map.permute(0, 2, 1).view(B, 2, H, H),
-                                                    size=args.img_size, mode='bilinear', align_corners=True)
-                        anomaly_map = torch.softmax(anomaly_map, dim=1)
-                        seg_loss += loss_focal(anomaly_map, mask)
-                        seg_loss += loss_dice(anomaly_map[:, 1, :, :], mask)
-                    
-                    loss = seg_loss + det_loss # = focal(seg_out, mask) + bce(det_out, y)
-                    loss.requires_grad_(True)
-                    seg_optimizer.zero_grad()
-                    det_optimizer.zero_grad()
-                    loss.backward()
-                    seg_optimizer.step()
-                    det_optimizer.step()
-
-                else:
-                    loss = det_loss
-                    loss.requires_grad_(True)
-                    det_optimizer.zero_grad()
-                    loss.backward()
-                    det_optimizer.step()
-
-                loss_list.append(loss.item())
-
-        train_dataset.shuffle_dataset()
-        train_loader = torch.utils.data.DataLoader(train_dataset, batch_size=1, shuffle=True, **kwargs)
-
-        # logs
-        print("Loss: ", np.mean(loss_list))
+    score = test(args, model, test_loader, text_feature_list[CLASS_INDEX[args.obj]])
         
-
 
 
 def test(args, seg_model, test_loader, text_features):
